updated validate middleware
