+++ application/controllers/todoController.js
@@ -1,6 +1,6 @@
 const todoService = require('../services/todoServices');
-const { validationResult } = require('express-validator');
-const { createTodoValidation } = require('../validation/todoValidation');
+// const { validationResult } = require('express-validator');
+// const { createTodoValidation } = require('../validation/todoValidation');
 
 const getAllTodos = async (req, res) => {
     try {
@@ -42,6 +42,20 @@ const getAllTodos = async (req, res) => {
 }
 
 const createTodo = async (req, res) => {
+
+    // //check for validation errors
+    // const errors = validationResult(req);
+    // if (!errors.isEmpty()) {
+    //     console.log(errors.array()[0].msg);
+    //     return res.status(422).json({
+    //         status: "error",
+    //         statusCode: 422,
+    //         message: "Validation errors",
+    //         // different way of error presenting
+    //         errors:errors.array()
+            
+    //     })
+    // }
     try { 
         const { title, description, dueDate, completed } = req.body;
         const newTodo = await todoService.createTodo(
@@ -142,10 +156,11 @@ const deleteTodo = async (req, res) => {
 
 module.exports = {  
     getAllTodos,
-    createTodo: [
-        createTodoValidation,
-        createTodo
-    ],
+    // createTodo: [
+    //     createTodoValidation,
+    //     createTodo
+    // ],
+    createTodo,
     updateTodo,
     deleteTodo
 }
